Replace main add_responsive_search_ad example with newer version that shows how to include geo targeting and image extensions.
